+++ Arduino/pid.h
@@ -1,11 +1,11 @@
 #pragma once
 typedef int WindupMode;
 
-void setPoint(float sp);
-void setProcessVariable(float pv, float time);
-float getManipulatedVariable();
+void setPoint(double sp);
+void setProcessVariable(double pv, double time);
+double getManipulatedVariable();
 
-const int EnableWhenControllable = 1, CapIntegralTerm = 2, EraseHistory = 3; 
+const int EnableWhenControllable = 1, CapIntegralTerm = 2, EraseHistory = 3;
 void setIntegralWindupMode(WindupMode mode);
 
 //Loop cycletime
